Removed a few foot-bot sensors, and made more generic implementations for them that the foot-bot uses. In particular:
footbot wheels actuator -> differential steering actuator
footbot leds actuator -> leds actuator
footbot proximity sensor -> ring proximity sensor

A new, general-purpose implementation of the proximity sensor was created. The ring proximity sensor is not yet done (it's there but not compiled) and will be done next. It will depend on the general-purpose proximity sensor.

For the time being these new devices will stay, to allow for thorough testing.
In a later stage they may be restored, although if things go as they should, there should be no need.

Changed the base definition of the simulated sensor and actuators to a leaner one.

Rename and moved a few files to reflect better the way things should be organized.