Added the proximity-sensor- and light-sensor-equipped entities. They store the positions of the corresponding sensors, and are used by the implementation of the corresponding sensors to calculate the final reaults.
Restored the footbot proximity sensor with the same interface as ARGoS2, and made its implementation as a decorator of the default proximity sensor implementation. The function used to calculate the final result is the correct one from ARGo2, but it's not normalized.

Removed the ring proximity sensor, as it is now useless.