Fixed bug in ray cast query of Dynamics2D engine.
The method CDynamics2DEngine::CheckIntersectionWithRay() was erroneously checking only the first hit reported by Chipmunk. When a short object (on the Z axis) was found by the engine, this object was the only one checked for 3D intersection. If this object was not a hit, no further object was checked, resulting in buggy compenetrations.