Changed order of operations in CSpace::Update().
Before Draw was executed after Act, making it difficult to see exactly on what sensor values the robot was about to act.

Now Act is performed first, then Sense and finally ControlStep. In this way, Draw displays the latest state available to the robot.

Commented out CSpaceMultiThreadBalanceLength because it does not work. Bug fixes will follow.