LuA controller is now capable of reading parameters from the configuration file and make them available to the lua script via the table robot.params.
The parameters are child tags of the params tag. The tag name is the key and the text between the tags becomes the values. The values are always passed as string to the lua controller. The type casting have to be done by the lua script. In the example below, the robot.params table has a key "test" with the value "Hello World". The keys can be arbitrary choosen.

<lua_controller id="red">
...
  <params script="lua_controller/red.lua">
    <test>Hello World</test>
  </params>
</lua_controller>