Solves #32.
The bug was due to the fact that physics engines were selecting the
closest occlusion. However, in many cases, the closest occlusion is the
robot that casts the ray to check (e.g., from its center). This, in
turn, invalidated the search for other occlusions, creating false
negatives.

The solution is to collect all the intersections from the physics
engines, and then to pick the closest in a dedicated function.

TODO:
- Duplicated intersections occur now, and it would be nice to filter
  them out. Possible solution: using a map.