generic: add a tag actuator
This commit adds a tag actuator that can set the payload of a tag. The actuator enables simulating tags which are displayed on small LCD or TFT screens and which are therefore
dynamic. The use case for such an actuator is similar to the LED actuator in that it enables situated communication but with a higher bandwidth than LEDs.