Add the OBJ and MTL files for drawing Spiri
This commit adds the 3D model for the Spiri quadcopter. Although the files for the model should be located under plugins/robots, they are located under plugins/simulator until we have a good solution to resource management in the simulator (see issue #94). The CMakeLists.txt has been modified so that the model directory is installed.