RFC initial attempt to improve error handling with Lua.
The error handling in CLuaController and CLuaUtility is a mixture of returning booleans, raising exceptions, and setting flags. This is an absolute mess and needs to be improved. This initial commit fixes one of the worst issues by reporting errors from Lua error during executing step/reset etc. Previously these errors were unreported by the simulator resulting in zombie robots.