Rename and update the Pi-Puck's front camera sensor
This commit updates the control interface and simulation implementation of the Pi-Puck's front camera. The terminology "camera system" has been dropped in favor of "front camera" since this will make it easier to accomodate 3rd party omnidirectional camera sensors. A simple test that determines whether a tag on a stigmergic block and a nearby LED can be detected has been added.